requirements.txt
apache-airflow==2.7.2
psycopg2==2.9.7
boto3==1.28.26
python-dotenv==1.0.0
apache-airflow-providers-amazon==8.8.0
apache-airflow-providers-postgres==5.7.1
dbt-postgres==1.7.1